meson: Add meson 0.57.1 recipe.
The current version of meson (0.53.2) is not compatible with
virglrenderer 1.1.0, which requires a newer version of meson
to build correctly. To address this issue, add meson 0.57.1
to the available versions.

To maintain backwards compatibility and avoid build breakage,
keep the default version of meson as 0.53.2. However, for
virglrenderer 1.1.0, set the PREFERRED_VERSION to 0.57.1 to
ensure that it builds with the required version of meson.

Origin of the import:
Repo: git://git.openembedded.org/openembedded-core
Branch: master
Revision: 958d7f8cebe863705dc6710b671764879ea68575
Directory: meta/recipes-devtools/meson

Change-Id: I4c5bfc0d5af307631bb22256a3cf43584cb0a0df
Signed-off-by: Pavlo Kandrin <pkandrin@qti.qualcomm.com>