recipes-kernel/edk2: Adapt EDK2 bootloader patches to the Snapdragon_Auto.LV.3.2.1 BSP.
The COQOSHV SDK and use case metalayers include patches to the EDK2
bootloader sources. The COQOSHV SDK is intended to be used with
SA8155P.HOS.3.1.3.1 and EDK2 revision 'auto-uefi.lnx.1.9.1.r3-rel'.

In the Snapdragon_Auto.LV.3.2.1 BSP the 'uefi.lnx.1.9.1.r15-rel'
revision is used hence the patches cannot be applied 'as is' and need to
be reworked.

Changes include:
 - Add support for using outer dtb in edk2 bootloader
 - Disable parallel load and verification feature in ABL
 - Store device lock state and serial number
 - Remove keymaster ROT related setup in native bootloader
 - Remove device lock state set in dedicated memory
 - Send blow sw fuse command to trustzone
 - Update DevInfo struct to share it between native and guest bootloaders
 - Provide boot tamper state to TZ if the boot state is ORANGE
 - Blow milestone fuse
 - Store recovery wipe command for GVM when device lock state change

Change-Id: I70ee346345314999786c7d7acb7c30a51a6dbf67
Signed-off-by: Ankurkumar Satasiya <asatasiy@qti.qualcomm.com>