linux-msm: Add patch to support early QDSS init.
QCPE now supports early initialization of QDSS / ARM coresight for
tracing. For this it has to talk to the Always On Processor (AOP)
through its mailbox. Unfortunately, the Linux driver for the AOP mailbox
fails on startup when a link has already been pre-established. This adds
a patch to the driver in 'msm_qmp.c' to fix that. It follows the patches
description:

   msm_qmp.c: Support a pre-established link.

   Support the case that a link has already been established before
   Linux boots up. Such an entity could be the bootloader or any part
   of the firmware. A concrete use-case for this is to bringup up the
   QDSS clocks early in order to initialize ARM coresight.

   To support this use-case the RX worker function is extended to check
   for an already established link in the 'LINK_DISCONNECTED' state.
   If the link exists we jump directly to 'LINK_CONNECTED' and signal
   completion.

Change-Id: Ia26f30b949fccbb20c2adcca94cef776925e78bf
Signed-off-by: Ulrich Meis <umeis@qti.qualcomm.com>