linux-msm: Fix processing CPU cluster topology updates.
This fix allows assigning any CPU core to Linux PVM in COQOS-enabled
environment. It iterates through all the CPU cores in a cluster instead
of taking only the first one. Kernel scheduler routines expect that CPU
cores in clusters are being enabled starting from the lowest index.
In virtual environment any CPU core has to be allowed to be assigned.

Change-Id: Ib384fba68ac627f23435a68ea3d13a86f9c744f7
Signed-off-by: Marcin Mozejko <mmozejko@qti.qualcomm.com>