linux-msm: pinctrl-msm: Don't call irq_chip_eoi_parent for interrupts routed via dirconn.
In COQOS hvpervisor every invalid EOI signal is documented as resulting
in an unpredictable behavior of the hardware. An invalid EOI preceding a
valid one causes the hardware to ignore the latter. The mitigation of
that issue is to skip calling irq_chip_eoi_parent() when interrupts are
routed via dirconn and thereby to avoid sending an invalid EOI.

Change-Id: I1ba79bce68bdd0eaf01abc05b6e7e0b2f0df0aa6
Signed-off-by: Marcin Mozejko <mmozejko@qti.qualcomm.com>