linux-msm: Add QCPE 3.0.0 support for HOS.
This change adds support for QCPE 3.0.0 to SCM driver. New design
utilizes COQOSHV channels, shared memory and interrupts for a more
efficient SMC handling. The new approach utilizes HVC instruction
to trigger the communication with QCPE and expects an interrupt
signalled response. In that way a VCPU is not being blocked because
of waiting for a SMC completion.

Change-Id: I4bf1536e9e911f2374ba01f5d3cc049d897bfacc
Signed-off-by: Marcin Mozejko <mmozejko@qti.qualcomm.com>