system-core: Disable CPU isolation for core7.
The init_post_boot script which is called by init_post_boot.service
sets the /sys/devices/system/cpu/cpu7/core_ctl/min_cpus to 0 which
enables the core isolation but due to a bug in the gpio direct connectd
interrupt the core7 is put into suspend with pending interrupts when
core isolation is enabled. This change will disable the core isolation
and hide the gpio interrupt bug.

Change-Id: I13f040eae922f01767dfb8725a6d544552aa2eaa
Signed-off-by: Anandu Sudhakaran <anansudh@qti.qualcomm.com>