systemd: Create console instances.
Generator automatically instantiates serial-getty@.service on the kernel
console(s), if they can function as ttys and are not provided by the
virtual console subsystem.

In particular it instantiates serial-getty@.service instances for both
native and/or virtualized console if it is available.

This change is needed to get access to virtualized Linux and Android
consoles.

Change-Id: I0dda23dbb31b958fe5176e860adc650c5e1f8397
Signed-off-by: Marcin Mozejko <mmozejko@qti.qualcomm.com>