recipes-security: packagegroup-core-selinux.bbappend: Remove dependency on 'selinux-autorelabel'.
Standard content of the 'packagegroup-core-selinux' includes a
'selinux-autorelabel' feature, which performs relabeling of all
mount points, including the root file system, during the first boot.
However, this feature uses an '.autorelabel' file in the file system
root as a marker to initiate relabeling. Due to the read-only file system,
this file cannot be deleted, causing the relabeling procedure to repeat
after reboot. Since this project configuration already includes labeling
of the root file system during build, the runtime 'selinux-autorelabel'
feature shall be disabled.

Change-Id: Ia80340b94f9d5c6f2c2316f6495b7fa13c9d62b9
Signed-off-by: Andrii Smolskyi <asmolsky@qti.qualcomm.com>