edk2_git.bbappend: Disable dtbo overlay.
PVM is not using dtbo partition but if dtbo is flashed, bootloader will
try to overlay the device tree with the content of dtbo image and it
will lead to undesired behavior of the guest(most likely the result
will be crash of the guest). So, explicitly disable the dtbo need from
bootloader.

Change-Id: Ide4bc9a9aa6711e777763a5f304aa6afa5c238fd
Signed-off-by: Ankurkumar Satasiya <asatasiy@qti.qualcomm.com>
(cherry picked from commit 594d271b3efba78885a84b21c912ab025d7fcb07)