recipes-kernel: Disable MSM GENI console driver.
Being enabled the CONFIG_SERIAL_MSM_GENI_CONSOLE option negatively
impacts the suspend-to-ram feature when the MSM GENI serial console
is not being used.

The GENI Serial Engine driver adds bus bandwidth votes during
initialization, but these votes are never removed. It is expected that
the MSM GENI console driver would remove these votes as part of its
operation. In the COQOS HV hosted system, a virtual serial port is used
for the console, so the votes are never removed, resulting in them
remaining active and blocking suspend-to-ram operations.

Change-Id: Ia42cb18fa21e6e5f35e8e2532e2455f6cb485294
Signed-off-by: Andrii Smolskyi <asmolsky@qti.qualcomm.com>