classes: Add coqos-qimage class.
coqos-qimage class changes the content of the vbmeta image. New vbmeta
reflects changes made to the LV during COQOS integration. DTBO image
descriptor is left in place despite the image is not being used. This is
because EDK2 bootloader requires DTBO to be present. So to keep it more
simple - DTBO is loaded and verified but not applied.

Change-Id: Ic30821ab689fce020a4d80823b73558948e10e36
Signed-off-by: Marcin Mozejko <mmozejko@qti.qualcomm.com>